[#189] MODIFY: Constant 파일 위치 변경
