--- HousUIComponent/Sources/Constant.swift
@@ -1,24 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by 김지현 on 2023/05/21.
-//
-
-import Foundation
-
-enum HousButton {
-
-  enum Onboarding {
-    static let next = "다음으로"
-    static let createRoom = "방 만들기"
-    static let leaveProfile = "탈퇴하기"
-  }
-
-  enum Login {
-    static let kakao = "카카오톡으로 계속하기"
-    static let apple = "Apple로 계속하기"
-    static let save = "저장하기"
-    static let retry = "다시 검사해보기"
-  }
-}
