--- HousUIComponent/Sources/HousButton/File.swift
@@ -1,19 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by 김지현 on 2023/05/21.
-//
-
-import UIKit
-
-internal final class OnboardingButton: UIButton {
-
-    internal override init(frame: CGRect) {
-        super.init(frame: .zero)
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
